adding new binaries and del folder scr
